refactor: consolidate Gemini API and clean up debug code
- Add FlashcardGenerator implementation to GeminiApiDataAccess
- Remove separate GeminiFlashcardGenerator class
- Remove debug print statements from CourseWorkspaceView
- Update AppBuilder to use shared geminiDAO
