+++ src/main/java/app/AppBuilder.java
@@ -1,8 +1,6 @@
 package app;
 
-import data_access.DemoCourseAccess;
 import data_access.GeminiApiDataAccess;
-import data_access.GeminiFlashcardGenerator;
 import data_access.LocalCourseRepository;
 import entities.Course;
 import entities.PDFFile;
@@ -72,7 +70,7 @@ public class AppBuilder {
 
     public AppBuilder addWriteTestView() {
         mockTestViewModel = new MockTestViewModel();
-        writeTestView = new WriteTestView(mockTestViewModel); // The view for taking the test
+        writeTestView = new WriteTestView(mockTestViewModel);
         cardPanel.add(writeTestView, mockTestViewModel.getViewName());
         return this;
     }
@@ -101,23 +99,13 @@ public class AppBuilder {
     }
 
     public AppBuilder addEvaluateTestUseCase() {
-        // The Presenter for the evaluation results view
         EvaluateTestPresenter evalPresenter = new EvaluateTestPresenter(evaluateTestViewModel, loadingViewModel, viewManagerModel);
-
-        // The Interactor for the evaluation use case. It correctly uses the DAOs.
         EvaluateTestInteractor evalInteractor = new EvaluateTestInteractor(courseDAO, geminiDAO, evalPresenter);
-
-        // The Controller that the WriteTestView will use to trigger the evaluation.
         EvaluateTestController evalController = new EvaluateTestController(evalInteractor);
-
-        // The Presenter for the WriteTestView's navigation (next/prev question).
         MockTestPresenter mockTestPresenter = new MockTestPresenter(mockTestViewModel, viewManagerModel, loadingViewModel);
 
-        // Inject both the controller (for submitting) and the presenter (for navigation) into the WriteTestView.
         writeTestView.setController(evalController);
         writeTestView.setPresenter(mockTestPresenter);
-
-        // Inject the presenter into the EvaluateTestView
         evaluateTestView.setPresenter(evalPresenter);
 
         return this;
@@ -155,7 +143,6 @@ public class AppBuilder {
     }
 
     public AppBuilder addCourseUseCases() {
-        // presenter for dashboard + navigation
         CourseDashboardOutputBoundary courseDashboardPresenter =
                 new CourseDashboardPresenter(
                         viewManagerModel,
@@ -164,7 +151,6 @@ public class AppBuilder {
                         courseCreateViewModel
                 );
 
-        // local course repository for your use cases
         ICourseRepository courseRepository = new LocalCourseRepository();
 
         CourseDashboardInputBoundary courseDashboardInteractor =
@@ -172,7 +158,6 @@ public class AppBuilder {
         CourseDashboardController courseDashboardController =
                 new CourseDashboardController(courseDashboardInteractor);
 
-        // presenter for workspace / edit views
         CourseWorkspaceOutputBoundary coursePresenter =
                 new CoursePresenter(
                         viewManagerModel,
@@ -185,7 +170,6 @@ public class AppBuilder {
                 new CourseWorkspaceInteractor(courseRepository, coursePresenter, courseDashboardPresenter);
         CourseController courseController = new CourseController(courseWorkspaceInteractor);
 
-        // hook controllers into your views
         this.courseDashboardView.setCourseDashboardController(courseDashboardController);
         this.courseDashboardView.setCourseWorkspaceController(courseController);
 
@@ -202,10 +186,6 @@ public class AppBuilder {
 
     // === FLASHCARD: Flashcard methods ===
 
-    /**
-     * Adds the flashcard generation and display views to the application.
-     * @return this AppBuilder for method chaining
-     */
     public AppBuilder addFlashcardViews() {
         this.flashcardViewModel = new FlashcardViewModel();
 
@@ -219,27 +199,16 @@ public class AppBuilder {
         return this;
     }
 
-    /**
-     * Wires up the flashcard generation use case with all necessary dependencies.
-     * @return this AppBuilder for method chaining
-     */
     public AppBuilder addFlashcardGenerationUseCase() {
-        // Create the flashcard generator (using Gemini API)
-        GeminiFlashcardGenerator generator = new GeminiFlashcardGenerator();
-
-        // Create the presenter
         GenerateFlashcardsPresenter presenter =
                 new GenerateFlashcardsPresenter(flashcardViewModel, viewManagerModel);
 
-        // Create the interactor
         GenerateFlashcardsInteractor interactor =
-                new GenerateFlashcardsInteractor(generator, presenter);
+                new GenerateFlashcardsInteractor(geminiDAO, presenter);
 
-        // Create the controller
         GenerateFlashcardsController controller =
                 new GenerateFlashcardsController(interactor);
 
-        // Inject controller into the view
         generateFlashcardsView.setController(controller);
         courseWorkspaceView.setFlashcardsController(controller);
 
@@ -252,8 +221,6 @@ public class AppBuilder {
         application.add(cardPanel);
         this.courseDashboardView.renderDashboard();
 
-
-        // Set the initial view
         viewManagerModel.setState(this.courseDashboardView.getViewName());
         viewManagerModel.firePropertyChange();
